+++ src-core/common/image/earth_curvature.cpp
@@ -32,7 +32,7 @@ namespace image
             Image<T> output_image(corrected_width, image.height(), image.channels()); // Allocate output image
 
             if (foward_table != nullptr)
-                for (int i = 0; i < image.width(); i++)
+                for (int i = 0; i < (int)image.width(); i++)
                     foward_table[i] = -1;
 
             for (int channel = 0; channel < image.channels(); channel++)
@@ -56,7 +56,7 @@ namespace image
             if (foward_table != nullptr)
             {
                 float last_val = 0;
-                for (int i = 0; i < image.width(); i++)
+                for (int i = 0; i < (int)image.width(); i++)
                 {
                     if (foward_table[i] == -1)
                         foward_table[i] = last_val;
